Update GET-less object handling.
This is ground work for a generalization that allows filtering results
in the client-side (e.g. SQL Users).

A fix to lines() was also required to make it work consistently across
files with and without newline in the last line. This fix also does
not matter anymore how many newlines the file has at the end (lines(N)
will ensure that will have *exactly* N empty lines after the content.

Change-Id: Iba987821c9e2614984e078f847dc86849e39c56b