refactored puppet resourceref into a separate property class like nestedobject
ResourceRefs are used to reference a different block in the Puppet
manifest and return a set attribute from that block.

ResourceRef properties are broken into two classes: a class that has the
desired value straight from the GCP API (*Api) and a class that has the name of
the Puppet block (*Catalog). Both classes return the desired value (either directly
in the first case, or after fetching the proper value from the correct
Puppet block). A base class assures that these two classes can have the
desired properties compared.

Change-Id: I876556f3b114d4d50a047cf3cdc2731eb2ed18af