Added better README documentation for nested objects and arrays of nested objects
This adds better docs for nested objects and arrays of nested objects.
All properties are at the same depth, but use either [] or / to denote
being a part of arrays or nested objects.

Note: This does not affect regular arrays. My rationale is that regular
arrays (arrays of strings, integers, etc.) should be well-explained by
the description already.

Change-Id: Id2b6fd54e072567a93f5d721945c1cd996f27fe2